Some changes to Spectral clustering step
Spectral clustering had an error. I computed the gaussian similarities, but the spectral clustering fit method recomputed them. I changed method = "precomputed" to fix this. I further changed the mean that is used for spectral clustering to 2000 km. The results are not very sensitive to this mean (between about 1100 and 4000 km).

For the spectral clustering, I changed the similarity matrix so that the distance and std of distance are squared (to make it more similar to a gaussian). This didn't really influence the results, but it's more accurate. I also changed it so that there is no gamma: just use a gaussian similarity matrix. Also don't use dist = 0 parts of the matrix for calculating the std.