Added code to live-downloaded the Valoroso 2013 dataset
Also modified area_plot.py to get it working again. This will only work locally on my computer, unless somebody changes the path to the topography file within that file. Also added code so that once the Zenodo url for focal mechanism data is available, we can type in that url into load_data.ipynb and it should be more or less good to go.